+++ src/ts/Managers/OperativeEventManager.ts
@@ -64,7 +64,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.START)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.START]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve();
         }
 
@@ -90,7 +90,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.FIRST_QUARTILE)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.FIRST_QUARTILE]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
 
@@ -107,7 +107,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.MIDPOINT)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.MIDPOINT]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
 
@@ -124,7 +124,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.THIRD_QUARTILE)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.THIRD_QUARTILE]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
 
@@ -141,7 +141,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.SKIP)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.SKIP]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
         adUnit.getCampaign().getSession().setEventSent(EventType.SKIP);
@@ -178,7 +178,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.VIEW)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.VIEW]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
         adUnit.getCampaign().getSession().setEventSent(EventType.VIEW);
@@ -194,7 +194,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.CLICK)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.CLICK]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
         adUnit.getCampaign().getSession().setEventSent(EventType.CLICK);
@@ -368,8 +368,11 @@ export class OperativeEventManager {
 
     private createVideoEventUrl(adUnit: AbstractAdUnit, type: string): string {
         const campaign = adUnit.getCampaign();
-        if((campaign instanceof PerformanceCampaign || campaign instanceof MRAIDCampaign) && campaign.getVideoEventUrl(type)) {
-            return campaign.getVideoEventUrl(type);
+        if(campaign instanceof PerformanceCampaign || campaign instanceof MRAIDCampaign) {
+            const url = campaign.getVideoEventUrl(type);
+            if(url) {
+                return url;
+            }
         }
         return [
             OperativeEventManager.VideoEventBaseUrl,
@@ -383,20 +386,29 @@ export class OperativeEventManager {
 
     private createClickEventUrl(adUnit: AbstractAdUnit): string {
         const campaign = adUnit.getCampaign();
-        let url: string;
-        if((campaign instanceof PerformanceCampaign || campaign instanceof MRAIDCampaign) && campaign.getClickUrl()) {
-            url = campaign.getClickUrl();
-        } else {
+        let url: string | undefined;
+        let parameters: any = {};
+
+        if(campaign instanceof PerformanceCampaign || campaign instanceof MRAIDCampaign) {
+            const clickUrl = campaign.getClickUrl();
+            if(clickUrl) {
+                parameters = { redirect: false };
+                url = clickUrl;
+            }
+        }
+        if(!url) {
             url = [
                 OperativeEventManager.ClickEventBaseUrl,
                 campaign.getId(),
                 'click',
                 campaign.getGamerId(),
             ].join('/');
+            parameters = {
+                gameId: this._clientInfo.getGameId(),
+                redirect: false
+            };
         }
-        return Url.addParameters(url, {
-            gameId: this._clientInfo.getGameId(),
-            redirect: false
-        });
+
+        return Url.addParameters(url, parameters);
     }
 }