+++ src/ts/Models/Campaign.ts
@@ -1,6 +1,8 @@
 import { Asset } from 'Models/Assets/Asset';
 import { ISchema, Model } from 'Models/Model';
 import { Session } from 'Models/Session';
+import { WebViewError } from 'Errors/WebViewError';
+import { Diagnostics } from 'Utilities/Diagnostics';
 
 export interface ICampaign {
     id: string;
@@ -90,5 +92,10 @@ export abstract class Campaign<T extends ICampaign = ICampaign> extends Model<T>
     public abstract getRequiredAssets(): Asset[];
     public abstract getOptionalAssets(): Asset[];
     public abstract isConnectionNeeded(): boolean;
+
+    protected handleError(error: WebViewError) {
+        Diagnostics.trigger('set_model_value_failed', error, this.getSession());
+        throw error;
+    }
 
 }